klient/machine: move mount sync interfaces to supervisor file
